Fix more regressions, impl jkGUISaveLoad

+++ df2_reimpl/src/General/Darray.c
@@ -73,11 +73,11 @@ int Darray_sub_520CB0(Darray *a1, int (__cdecl *a2)(int, int), int a3)
     int v3; // edi
 
     v3 = 0;
-    if ( (signed int)a1->total <= 0 )
+    if (a1->total <= 0)
         return -1;
     while ( a2(a3, (int)Darray_GetIndex(a1, v3)) )
     {
-        if ( ++v3 >= (signed int)a1->total )
+        if ( ++v3 >= a1->total )
             return -1;
     }
     return v3;
